Uploading images for create and deleting images for update added
when a user creates a listing it uploads the image in the uploads folder then if the user edits their photos in update then it deletes the old image and replaces it with an new one. I also worked on the feature that adds cleanliness in out form which is the dropdown checkbox. i had to brush up on my JS to do the feature but now our form looks cleaner